# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "API мини-приложения реверс-найма в MAX.\nАвторизация: POST /api/v1/auth c initData из MAX Bridge,\nдалее заголовок Authorization: Bearer \u003ctoken\u003e (JWT, 7 дней).",
        "title": "MAX Mini App — Reverse Hiring API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/",
    "paths": {
...
        "/api/v1/companies": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company"
                ],
                "summary": "Создать компанию",
                "parameters": [
                    {
                        "description": "компания + позиция создателя",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.CompanyInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.Company"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/companies/{id}/vacancies": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vacancy"
                ],
                "summary": "Вакансии компании",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID компании",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/max-miniapp_backend_internal_dto.Vacancy"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "vacancy"
                ],
                "summary": "Создать вакансию",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID компании",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "вакансия",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.VacancyInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.Vacancy"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/v1/companies/{id}/verify": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Проверяет токен бота через GET /me платформы MAX и помечает компанию верифицированной.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company"
                ],
                "summary": "Верифицировать компанию",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID компании",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "токен бота",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.VerifyRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.Company"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/api/v1/my/companies": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "company"
                ],
                "summary": "Мои компании",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/max-miniapp_backend_internal_dto.Company"
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/max-miniapp_backend_internal_dto.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "max-miniapp_backend_internal_dto.Company": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Москва, ул. Тверская, 1"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "ИТ-компания"
                },
                "id": {
                    "type": "integer"
                },
                "logo_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "Ромашка"
                },
                "position": {
                    "type": "string",
                    "enum": [
                        "owner",
                        "hr",
                        "employee"
                    ],
                    "example": "owner"
                },
                "updated_at": {
                    "type": "string"
                },
                "verified": {
                    "type": "boolean"
                },
                "website": {
                    "type": "string",
                    "example": "https://romashka.ru"
                }
            }
        },
        "max-miniapp_backend_internal_dto.CompanyInput": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string",
                    "example": "Москва, ул. Тверская, 1"
                },
                "description": {
                    "type": "string",
                    "example": "ИТ-компания"
                },
                "logo_url": {
                    "type": "string"
                },
                "name": {
                    "type": "string",
                    "example": "Ромашка"
                },
                "position": {
                    "type": "string",
                    "enum": [
                        "owner",
                        "hr",
                        "employee"
                    ],
                    "example": "owner"
                },
                "website": {
                    "type": "string",
                    "example": "https://romashka.ru"
                }
            }
        },
...
        "max-miniapp_backend_internal_dto.ErrorResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "invalid request body"
                }
            }
        },
...
        "max-miniapp_backend_internal_dto.Vacancy": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "example": "Москва"
                },
                "company_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "description": {
                    "type": "string",
                    "example": "Платёжный сервис"
                },
                "employment_type": {
                    "type": "string",
                    "enum": [
                        "full_time",
                        "part_time",
                        "contract",
                        "internship"
                    ],
                    "example": "full_time"
                },
                "id": {
                    "type": "integer"
                },
                "is_active": {
                    "type": "boolean"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "min_experience_months": {
                    "type": "integer",
                    "example": 12
                },
                "required_skills": {
                    "type": "string",
                    "example": "go, postgres"
                },
                "response_ttl_hours": {
                    "type": "integer",
                    "example": 48
                },
                "salary_max": {
                    "type": "integer",
                    "example": 280000
                },
                "salary_min": {
                    "type": "integer",
                    "example": 180000
                },
                "title": {
                    "type": "string",
                    "example": "Go-разработчик"
                },
                "updated_at": {
                    "type": "string"
                },
                "work_format": {
                    "type": "string",
                    "enum": [
                        "onsite",
                        "hybrid",
                        "remote"
                    ],
                    "example": "hybrid"
                }
            }
        },
        "max-miniapp_backend_internal_dto.VacancyInput": {
            "type": "object",
            "properties": {
                "city": {
                    "type": "string",
                    "example": "Москва"
                },
                "description": {
                    "type": "string",
                    "example": "Платёжный сервис"
                },
                "employment_type": {
                    "type": "string",
                    "enum": [
                        "full_time",
                        "part_time",
                        "contract",
                        "internship"
                    ],
                    "example": "full_time"
                },
                "lat": {
                    "type": "number"
                },
                "lng": {
                    "type": "number"
                },
                "min_experience_months": {
                    "type": "integer",
                    "example": 12
                },
                "required_skills": {
                    "type": "string",
                    "example": "go, postgres"
                },
                "response_ttl_hours": {
                    "type": "integer",
                    "example": 48
                },
                "salary_max": {
                    "type": "integer",
                    "example": 280000
                },
                "salary_min": {
                    "type": "integer",
                    "example": 180000
                },
                "title": {
                    "type": "string",
                    "example": "Go-разработчик"
                },
                "work_format": {
                    "type": "string",
                    "enum": [
                        "onsite",
                        "hybrid",
                        "remote"
                    ],
                    "example": "hybrid"
                }
            }
        },
...
        "max-miniapp_backend_internal_dto.VerifyRequest": {
            "type": "object",
            "properties": {
                "bot_token": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Значение: \"Bearer \u003cjwt\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}